+++ tests/Dummies/Jobs/Foo.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Madewithlove\Tactician\Dummies\Jobs;
+
+class Foo
+{
+
+}