+++ tests/Dummies/Listeners/Foo.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Madewithlove\Tactician\Dummies\Listeners;
+
+class Foo
+{
+
+}