+++ tests/Dummies/Nested/Listeners/Foo.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Madewithlove\Tactician\Dummies\Nested\Listeners;
+
+class Foo
+{
+
+}