+++ tests/TestCase.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Madewithlove\Tactician;
+
+use PHPUnit_Framework_TestCase;
+use Mockery;
+
+abstract class TestCase extends PHPUnit_Framework_TestCase
+{
+    protected function tearDown()
+    {
+        Mockery::close();
+    }
+}