+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # CHANGELOG
 
 ## Unreleased
+
+- Added `Illuminate\Container` locator